Faster computation of modular polynomial roots
A descent method is applied to eliminate the field extension
used to define the f-invariant. When full elimination is performed,
this is equivalent to solving the classical modular polynomial equation
for the j-invariant.

This is up to 10-20x faster than solving the original equation
especially if the f-invariant is in a large field extension.