Faster and simpler Graeffe transforms (adams_operator)
The correct way to apply a Graeffe transform
is to apply the well-known classical formulas.

This is 2x-10x faster and uses less memory because each multiplication
has the same degree as the input polynomial.

This avoids a memory usage blowup especially when handling
degree 1000 polynomials over GF(p^24), when p is large
(more than 1000 bits).